extract Gf2_192 and Gf2_192Poly into separatate modules;
