+++ ergotree-interpreter/src/sigma_protocol.rs
@@ -11,6 +11,8 @@ mod crypto_utils;
 mod dht_protocol;
 mod dlog_protocol;
 mod fiat_shamir;
+mod gf2_192;
+mod gf2_192poly;
 mod proof_tree;
 mod sig_serializer;
 mod unchecked_tree;
