+++ ergotree-interpreter/src/sigma_protocol/prover.rs
@@ -10,11 +10,11 @@ use crate::sigma_protocol::crypto_utils::secure_random_bytes;
 use crate::sigma_protocol::dht_protocol;
 use crate::sigma_protocol::fiat_shamir::fiat_shamir_hash_fn;
 use crate::sigma_protocol::fiat_shamir::fiat_shamir_tree_to_bytes;
+use crate::sigma_protocol::gf2_192poly::Gf2_192Poly;
 use crate::sigma_protocol::proof_tree::ProofTree;
 use crate::sigma_protocol::unchecked_tree::UncheckedDhTuple;
 use crate::sigma_protocol::unproven_tree::CandUnproven;
 use crate::sigma_protocol::unproven_tree::CorUnproven;
-use crate::sigma_protocol::unproven_tree::Gf2_192Poly;
 use crate::sigma_protocol::unproven_tree::NodePosition;
 use crate::sigma_protocol::unproven_tree::UnprovenDhTuple;
 use crate::sigma_protocol::Challenge;
