+++ ergotree-interpreter/src/sigma_protocol/unproven_tree.rs
@@ -1,9 +1,7 @@
 //! Unproven tree types
 
-// TODO: remove after all todo! are implemented
-#![allow(clippy::todo)]
-
 use super::dht_protocol::FirstDhTupleProverMessage;
+use super::gf2_192poly::Gf2_192Poly;
 use super::proof_tree::ConjectureType;
 use super::proof_tree::ProofTree;
 use super::proof_tree::ProofTreeConjecture;
@@ -494,8 +492,11 @@ impl CthresholdUnproven {
         Self { children, ..self }
     }
 
-    pub(crate) fn with_polynomial(&self, _q: Gf2_192Poly) -> Self {
-        todo!()
+    pub(crate) fn with_polynomial(self, q: Gf2_192Poly) -> Self {
+        Self {
+            polinomial_opt: Some(q),
+            ..self
+        }
     }
 
     fn with_position(self, updated: NodePosition) -> Self {
@@ -518,47 +519,5 @@ impl CthresholdUnproven {
 
     pub(crate) fn is_real(&self) -> bool {
         !self.simulated
-    }
-}
-
-// TODO: extract
-#[derive(PartialEq, Debug, Clone)]
-pub struct Gf2_192Poly {}
-
-impl Gf2_192Poly {
-    pub(crate) fn from_byte_array(_challenge: Challenge, _: Vec<u8>) -> Self {
-        todo!()
-    }
-
-    pub(crate) fn evaluate(&self, _idx: usize) -> Gf2_192 {
-        todo!()
-    }
-
-    pub(crate) fn interpolate(
-        _points: Vec<u8>,
-        _values: Vec<Gf2_192>,
-        _value_at_zero: Gf2_192,
-    ) -> Gf2_192Poly {
-        todo!()
-    }
-
-    pub(crate) fn to_bytes(&self) -> Vec<u8> {
-        todo!()
-    }
-}
-
-// TODO: extract
-#[derive(PartialEq, Debug, Clone)]
-pub(crate) struct Gf2_192 {}
-
-impl From<Gf2_192> for Challenge {
-    fn from(_: Gf2_192) -> Self {
-        todo!()
-    }
-}
-
-impl From<Challenge> for Gf2_192 {
-    fn from(_: Challenge) -> Self {
-        todo!()
     }
 }
